services/extractor/requirements.txt
# ── Extractor base ────────────────────────────────────────
playwright
mysql-connector-python
requests
python-dotenv
yt-dlp

# ── API REST (reemplaza al form_server simple) ─────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.0

# ── Autenticación Microsoft (validar tokens JWT) ───────────
python-jose[cryptography]==3.3.0
httpx==0.27.2

# ── Processor (integrado) ─────────────────────────────────
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
groq==0.9.0
